AVRO-3616: C++ Fix compilation warnings (#1836)
* AVRO-3616: [C++]: Fix compilation warnings

Replace <boost/test/included/unit_test_framework.hpp> include with <boost/test/included/unit_test.hpp>


* AVRO-3616: [C++]: Fix compilation warnings

Fix the order of constructor parameters


* AVRO-3616: Enable -pedantic and -Werror for compiler flags

Those should prevent from introducing new warnings.

-Wextra reports "unused method arguments" and those could not be fixed
without API break.


* AVRO-3616: Use size_t consistently for node->leaves()


(cherry picked from commit d336c6c7b91f9f39de2a5e34ae55c1d7c118507b)

+++ lang/c++/api/NodeImpl.hh
@@ -544,9 +544,9 @@ NodeImpl<A, B, C, D, E>::printBasicInfo(std::ostream &os) const {
         os << " " << sizeAttribute_.get();
     }
     os << '\n';
-    int count = leaves();
+    size_t count = leaves();
     count = count ? count : names();
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
         if (C::hasAttribute) {
             os << "name " << nameAt(i) << '\n';
         }
